updated models for using non static functions

+++ src/Models/GroupUser.php
@@ -7,6 +7,7 @@
  */
 
 namespace CoffeeRunner\Models;
+use CoffeeRunner\Utilities\DatabaseConnection;
 
 
 #TODO: deleteGroupUser should check both groupID and userID
@@ -32,12 +33,57 @@ class GroupUser implements \JsonSerializable
         );
         return $rtn;
     }
-
+//---------------------------Check this again---------------------------------
     public function createGroupUser()
     {
         //TODO: creates an entry into the db, use groupUserID, groupID, userID
-    }
+        try
+        {
+            $dbh = DatabaseConnection::getInstance();
+            $stmtHandle = $dbh->prepare(
+                "INSERT INTO 'groupUser'(
+                'groupID',
+                'userID')
+                VALUES (:groupID,:userID)");
+            $stmtHandle->bindValue(":groupID",$this->groupID);
+            $stmtHandle->bindValue(":userID",$this->userID);
 
+            $success = $stmtHandle->execute();
+
+            if(!$success)
+            {
+                throw new \PDOException("sql query execution failed");
+            }
+            $id = $dbh->lastInsertId();
+            return $this->getGroupUser($id);
+        }
+        catch(\Exception $e)
+        {
+            throw $e;
+        }
+
+    }
+    public function getGroupUser()
+    {
+        try
+        {
+            $dbh = DatabaseConnection::getInstance();
+            $stmtHandle = $dbh->prepare(
+                "SELECT * FROM 'groupUser' WHERE 'groupUserID'= :groupUserID");
+            $stmtHandle->bindValue(":groupUserID", $this->groupUserID);
+            $success = $stmtHandle->execute();
+            if(!$success){
+                throw new \PDOException("sql query execution failed");
+            }
+            $groupUser = $stmtHandle->fetch(\PDO::FETCH_CLASS,"CoffeeRunner/Models/GroupUser");
+            return $groupUser;
+        }
+        catch(\Exception $e)
+        {
+            throw $e;
+        }
+    }
+//---------------------------Check this again---------------------------------
     public function deleteGroupUser()
     {
         try
@@ -49,8 +95,9 @@ class GroupUser implements \JsonSerializable
             else
             {
                 $dbh = DatabaseConnection::getInstance();
-                $stmtHandle = $dbh->prepare("DELETE FROM `groupUser` WHERE `groupUserID` = :groupUserID");
-                $stmtHandle->bindValue(":groupUserID", $this->getGroupUserID());
+                $stmtHandle = $dbh->prepare("DELETE FROM `groupUser` WHERE `groupUserID` = :groupUserID AND 'userID' = :userID");
+                $stmtHandle->bindValue(":groupUserID", $this->groupUserID);
+                $stmtHandle->bindValue(":userID", $this->userID);
                 $success = $stmtHandle->execute();
 
                 if (!$success) {
@@ -83,10 +130,6 @@ class GroupUser implements \JsonSerializable
     /**
      * @param mixed $groupID
      */
-    public function setGroupID($groupID)
-    {
-        $this->groupID = $groupID;
-    }
 
     /**
      * @return mixed
@@ -101,6 +144,7 @@ class GroupUser implements \JsonSerializable
      */
     public function setUserID($userID)
     {
+        $userID = filter_var($userID,FILTER_SANITIZE_STRING);
         $this->userID = $userID;
     }
 }
